Resolve merge conflict.
